refactor: split off the git functions
